GyaanSetu/requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
setuptools==71.0.3
sqlparse==0.5.1
wheel==0.43.0
